Refactor underlined anchor on hover
So, after the previous changes I noticed I'd broken the underline on hover.
Upon closer inspection and research, I discovered that what I'd done was
actually a little inaccessible. It doesn't work on IE (boo) and it was
behaving really inconsistently. So, I removed the :root file declaring
font variables. This is mostly because I wasn't confident in the
implementation and I want to do some more research on it. I changed
the text-decoration into a border-bottom and I added some white drop
shadows to still accent the text through the underline itself.
This is working a lot better.