Refactor to SolidJS
My version of Gatsby was well out of date and I didn't have much of an appetite for using the latest
version of Gatsby. Instead, I've converted to SolidJS running within a Docker container as a way of
getting my head around Docker and getting more familiar with the tooling.